+++ php/Blockonomics.php
@@ -840,7 +840,7 @@ class Blockonomics
     public function check_paid_amount($paid_satoshi, $order, $wc_order){
         $order['paid_satoshi'] = $paid_satoshi;
         $paid_amount_ratio = $paid_satoshi/$order['expected_satoshi'];
-        $order['paid_fiat'] =number_format($order['expected_fiat']*$paid_amount_ratio,2,'.','');
+        $order['paid_fiat'] =number_format($order['expected_fiat']*$paid_amount_ratio,wc_get_price_decimals(),'.','');
 
         // This is to update the order table before we send an email on failed and confirmed state
         // So that the updated data is used to build the email