+++ webpack.config.js
@@ -1,0 +1,21 @@
+const defaultConfig = require('@wordpress/scripts/config/webpack.config');
+const WooCommerceDependencyExtractionWebpackPlugin = require('@woocommerce/dependency-extraction-webpack-plugin');
+const path = require('path');
+
+
+// Export configuration.
+module.exports = {
+	...defaultConfig,
+	entry: './js/block.js',
+	output: {
+		path: path.resolve( __dirname, 'build' ),
+		filename: 'block.js',
+	},
+	plugins: [
+		...defaultConfig.plugins.filter(
+			(plugin) =>
+				plugin.constructor.name !== 'DependencyExtractionWebpackPlugin'
+		),
+		new WooCommerceDependencyExtractionWebpackPlugin()
+	]
+};